third_party/blink/web_tests/fast/storage/storage-disallowed-in-data-url-expected.txt

This tests that storage methods throw exceptions with reasonable messages inside 'data:' URLs.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS exceptionMessage is "Failed to read the 'localStorage' property from 'Window': Storage is disabled inside 'data:' URLs."
PASS exceptionMessage is "Failed to read the 'sessionStorage' property from 'Window': Storage is disabled inside 'data:' URLs."
PASS successfullyParsed is true

TEST COMPLETE
